+++ src/clld/web/adapters/base.py
@@ -28,6 +28,10 @@ class Renderable:
         self.obj = obj
 
     @property
+    def label(self):
+        return getattr(self, '__label__', self.__class__.__name__)
+
+    @property
     def charset(self):
         return 'utf-8' \
             if self.mimetype.startswith('text/') \