+++ tests/test_web_adapters.py
@@ -2,11 +2,18 @@ from clld.interfaces import IIndex, IRepresentation
 from clld.db.models.common import Contribution, Language, Dataset
 
 
-def test_BibTex(env):
+def test_BibTex(env, env_factory):
     from clld.web.adapters import BibTex
 
     adapter = BibTex(None)
     assert '@' in adapter.render(Contribution.first(), env['request'])
+
+    res = adapter.render(Dataset.first(), env_factory(zenodo_concept_doi='xyz')['request'])
+    assert 'xyz' in res
+
+    env_ = env_factory(zenodo_version_doi='xyz', zenodo_version_tag='v1.0')
+    res = adapter.render(Dataset.first(), env_['request'])
+    assert 'xyz' in res and 'v1.0' in res
 
 
 def test_TxtCitation(env):
@@ -14,6 +21,7 @@ def test_TxtCitation(env):
 
     adapter = TxtCitation(None)
     assert '.' in adapter.render(Contribution.first(), env['request'])
+    assert adapter.label
     adapter.render(Dataset.first(), env['request'])